Fix BFloat16 inference error by casting output to float32 before numpy
When mixed precision (BF16) is enabled, model outputs remain in
bfloat16 dtype which numpy cannot handle directly. Adding .float()
before .numpy() ensures the conversion always succeeds.

+++ python_server/dlclassifier_server/services/inference_service.py
@@ -502,7 +502,7 @@ class InferenceService:
                             outputs = model(batch_tensor)
                     else:
                         outputs = model(batch_tensor)
-                    batch_logits = outputs.cpu().numpy()
+                    batch_logits = outputs.cpu().float().numpy()
 
             # Softmax per-tile and collect
             for i in range(batch_logits.shape[0]):
